extract dual panel library screen logic
